Refactored is-record-status and assert-record-status helpers
