--- packages/suspense/src/utils/isPendingRecord.ts
@@ -1,8 +1,0 @@
-import { STATUS_PENDING } from "../constants";
-import { PendingRecord, Record } from "../types";
-
-export function isPendingRecord(
-  record: Record<any>
-): record is PendingRecord<any> {
-  return record.status === STATUS_PENDING;
-}
